back-end/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1

langchain>=1.2.0
langchain-core>=1.2.0
langchain-community>=0.4.0
langchain-google-genai>=4.0.0
langchain-huggingface>=0.1.0
langchain-qdrant>=1.0.0
langgraph>=1.1.0
langgraph-checkpoint>=2.0.0

qdrant-client==1.11.2
rank-bm25==0.2.2
sentence-transformers==3.0.1
trafilatura==1.12.2
httpx==0.27.2
prometheus-fastapi-instrumentator==7.1.0
pymongo>=4.6.0
